% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablest.R
\name{fPrint}
\alias{fPrint}
\title{Print formated number depents on the number size}
\usage{
fPrint(z, digits = getOption("digits"), width = NULL, big.mark = ",")
}
\arguments{
\item{z}{a number.}

\item{digits}{the maximum number of digits to the right of
the decimal point}

\item{width}{\code{default} method: the \emph{minimum} field width or
    \code{NULL} or \code{0} for no restriction.

    \code{AsIs} method: the \emph{maximum} field width for non-character
    objects.  \code{NULL} corresponds to the default \code{12}.
  }

\item{big.mark}{%
    used for prettying (longish) numerical and complex sequences.
    Passed to \code{\link[base]{prettyNum}}: that help page explains the details.}
}
\value{
A formated character representations of \code{z}
}
\description{
The \code{fPrint} function format number depends on its own size.
}
\examples{
fPrint(NA)
fPrint(0)
fPrint(0, digits = 3)
fPrint(0.12345)
fPrint(0.1)
fPrint(0.012345)
fPrint(1.121234)
fPrint(11.121234)
fPrint(111.12123)
fPrint(11111)
}
