% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablest.R
\name{fStringMatch}
\alias{fStringMatch}
\title{Match}
\usage{
fStringMatch(x, sub_x, method = "exact")
}
\arguments{
\item{x}{A character vector.}

\item{sub_x}{A character vector.}

\item{method}{A string, used to set matching method, default is "exact".}
}
\value{
The position \code{sub_x} in \code{x}
}
\description{
Match
}
\examples{
x <- c("A", "B", "C", "D", "A")
subx <- c("A", "C")
fStringMatch(x, subx)
}
