% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablest.R
\name{fZipChar}
\alias{fZipChar}
\title{zip character vector}
\usage{
fZipChar(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A named integer vector, the name represents elements if \code{x}
    which is different from former elements, the number represents number  
    of consecutive occurrences.
}
\description{
zip character vector
}
\examples{
x <- c("A", "B", "B", "D", "D", "A", "A" ,"E")
fZipChar(x)
}
