% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablest.R
\name{nobs.felm}
\alias{nobs.felm}
\title{Get number of felm regression observations}
\usage{
\method{nobs}{felm}(object)
}
\arguments{
\item{object}{An object created by lfe::felm}
}
\value{
An integer
}
\description{
Get number of felm regression observations
}
