% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabreg.R
\name{tabreg}
\alias{tabreg}
\title{Tabulating Regresssion for Report and Further Processing}
\usage{
tabreg(
  ...,
  reglist = NULL,
  caption = NULL,
  outfmt = "text",
  align = NULL,
  vari = list(name = NULL, label = NULL),
  esti = list(estimate = 3L, std.error = "(3)", statistic = NULL, p.value = NULL,
    singlerow = FALSE, fun = NULL, fun.args = NULL),
  star = list(cut = c(0.1, 0.05, 0.01), symbol = c("*", "**", "***")),
  stat = list(name = c("N", "r2"), label = c("N", "*R*^2^")),
  header = list(name = c("dep", "reg", "no")),
  note = TRUE,
  outfun = NULL,
  outargs = list(),
  outstyle = NULL
)
}
\arguments{
\item{reglist}{a list of regression results.}

\item{caption}{the table caption.}

\item{outfmt}{a string for output format, which can be \code{text}(default),
\code{kable}, \code{flextable}, \code{markdown}, or \code{pandoc}.}

\item{vari}{a named list indicating variable name and label. If \code{name} is \code{NULL},
all regressors are used. If \code{label} is \code{NULL}, variable names are used.
\code{name} needed to be a character vector, while \code{label} can be a character
vector or named list. If \code{label} is a vector, the length must equal
to the number of uesd variables.}

\item{esti}{a named list indicating which data to extract and in what
way to present them.}
}
\description{
Convert regress model results to a \code{data.frame}, \code{kable}, \code{flextable},
or other similar object. The output object can be printed directly to pdf,
html or docx file. \code{tabreg} provides some common settings.
The output object can be further precessed, such as adding lines and
footnote, merging cells, adjusting output format, etc. In additional,
the way of procesing model results and output format can also be customised.
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)
lm.D90 <- lm(weight ~ group - 1) # omitting intercept
clotting <- data.frame(
   u = c(5,10,15,20,30,40,60,80,100),
   lot1 = c(118,58,42,35,27,25,21,19,18),
   lot2 = c(69,35,26,21,18,16,13,12,12))
glm.1 <- glm(lot1 ~ log(u), data = clotting, family = Gamma)
glm.2 <- glm(lot2 ~ log(u), data = clotting, family = Gamma)
l.reg <- list(lm.D9, lm.D90, glm.1, glm.2)
tabreg(l.reg)
}
